move callback routes to global router
